# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TRADE FAST GOLD Investment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /investments:
    get:
      operationId: listInvestments
      tags: [investments]
      summary: List user investments
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, completed, cancelled]
      responses:
        "200":
          description: List of investments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Investment"
    post:
      operationId: createInvestment
      tags: [investments]
      summary: Create a new investment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvestmentInput"
      responses:
        "201":
          description: Created investment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investment"
  /investments/{id}:
    get:
      operationId: getInvestment
      tags: [investments]
      summary: Get investment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Investment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investment"

  # ─── WALLETS ──────────────────────────────────────────────────────────────
...
  /deposits:
    get:
      operationId: listDeposits
      tags: [deposits]
      summary: List user deposits
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, rejected]
      responses:
        "200":
          description: List of deposits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deposit"
    post:
      operationId: createDeposit
      tags: [deposits]
      summary: Create deposit request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositInput"
      responses:
        "201":
          description: Created deposit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"
  /deposits/{id}:
    get:
      operationId: getDeposit
      tags: [deposits]
      summary: Get deposit by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deposit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"

  # ─── WITHDRAWALS ──────────────────────────────────────────────────────────
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: kycStatus
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserPage"
  /admin/users/{id}:
    get:
      operationId: getAdminUser
      tags: [admin]
      summary: Get user detail (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
    patch:
      operationId: updateAdminUser
      tags: [admin]
      summary: Update user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
  /admin/deposits:
    get:
      operationId: listAdminDeposits
      tags: [admin]
      summary: List all deposits (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Paginated deposits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDepositPage"
  /admin/deposits/{id}/approve:
    patch:
      operationId: approveDeposit
      tags: [admin]
      summary: Approve deposit (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved deposit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"
  /admin/deposits/{id}/reject:
    patch:
      operationId: rejectDeposit
      tags: [admin]
      summary: Reject deposit (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: Rejected deposit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, clerkId, email, firstName, lastName, role, kycStatus, referralCode, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        kycStatus:
          type: string
          enum: [not_submitted, pending, approved, rejected]
        referralCode:
          type: string
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
...
    Investment:
      type: object
      required: [id, userId, planId, planName, amount, roiPercent, durationDays, status, startDate, endDate, profitEarned, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        planId:
          type: integer
        planName:
          type: string
        amount:
          type: number
        roiPercent:
          type: number
        durationDays:
          type: integer
        status:
          type: string
          enum: [active, completed, cancelled]
        startDate:
          type: string
        endDate:
          type: string
        profitEarned:
          type: number
        progressPercent:
          type: number
        createdAt:
          type: string
    InvestmentInput:
      type: object
      required: [planId, amount]
      properties:
        planId:
          type: integer
        amount:
          type: number

    # ─── WALLETS ────────────────────────────────────────────────────────────
    Wallet:
      type: object
      required: [id, userId, type, balance, currency, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [main, profit, bonus, referral]
        balance:
          type: number
        currency:
          type: string
        createdAt:
          type: string
...
    Deposit:
      type: object
      required: [id, userId, amount, method, status, reference, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
        method:
          type: string
          enum: [bank_transfer, card, bitcoin, ethereum, usdt_trc20, usdt_erc20]
        status:
          type: string
          enum: [pending, confirmed, rejected]
        reference:
          type: string
        transactionHash:
          type: ["string", "null"]
        proofUrl:
          type: ["string", "null"]
        rejectionReason:
          type: ["string", "null"]
        confirmedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    DepositInput:
      type: object
      required: [amount, method]
      properties:
        amount:
          type: number
        method:
          type: string
          enum: [bank_transfer, card, bitcoin, ethereum, usdt_trc20, usdt_erc20]
        transactionHash:
          type: string
        proofUrl:
          type: string

    # ─── WITHDRAWALS ────────────────────────────────────────────────────────
...
    RejectInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string

    # ─── ADMIN ──────────────────────────────────────────────────────────────
...
    AdminUserPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    AdminUserDetail:
      type: object
      required: [user, wallets, totalInvested, totalProfit, totalDeposited, totalWithdrawn]
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"
        wallets:
          type: array
          items:
            $ref: "#/components/schemas/Wallet"
        totalInvested:
          type: number
        totalProfit:
          type: number
        totalDeposited:
          type: number
        totalWithdrawn:
          type: number
    AdminUserUpdate:
      type: object
      properties:
        isActive:
          type: boolean
        role:
          type: string
          enum: [user, admin]
        kycStatus:
          type: string
          enum: [not_submitted, pending, approved, rejected]
    AdminDepositPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Deposit"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...